requirements.txt
google_adk==1.9.0
mysql-connector-python
pymysql
python-dotenv
nest-asyncio
google-generativeai
Deprecated
agentops